requirements.txt
Flask>=3.0,<4
requests>=2.31
python-dotenv>=1.0
gunicorn>=21
